backend/requirements.txt
Django>=4.2.3,<=4.3
djangorestframework>=3.14.0,<=3.15
opencv-python>=4.8.0,<=4.9
PyJWT>=1.7.1,<=1.8